Update README.md
Windows 10 IoT Core is out of support since November 2020. (see https://docs.microsoft.com/en-us/lifecycle/products/windows-10-iot-core)

By having the Broadcom GENET Ethernet driver built with the NetAdapterCx driver framework, which is only available on Windows 10 version 2004 onwards, this repository is not compilable as a whole for earlier releases. The last release of Windows 10 IoT Core to exist is version 1809.

Should the CoC section remain as it's no longer under the Microsoft umbrella but on the Raspberry Pi org?