bcmemmc2: use OverrrideFlags instead of driver-level flags (#29)
We're currently using service parameters (which apply at the driver
level) to disable UHS. We should instead be using OverrideFlags (which
apply at the device level).

- Disabled UHS since sdbus.sys can't configure the voltage regulator.
- Disabled 8BIT because SDHOST capabilities incorrectly claim 8-bit support.

For future consideration, sdbus.sys DMA works fine on RPi400. If we
could get a distinct hardware ID for the new chip (e.g. a REV), we could
enable DMA support on RPi400.